+++ src/main/java/eu/senla/client/database/DatabaseConnection.java
@@ -1,4 +1,4 @@
-package eu.senla.client;
+package eu.senla.client.database;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
